fix(steam): align shortcut scan diagnostics
